added range checks

+++ Debit_Cards_Project/DAL/Repositories/DebitCardRepository.cs
@@ -7,6 +7,7 @@ namespace Debit_Cards_Project.DAL.Repositories
     public sealed class DebitCardRepository : IDebitCardRepository
     {
         private readonly DebitCardsDB _db;
+        private int _length;
         private readonly ILogger<DebitCardRepository> _logger;
 
             public DebitCardRepository(DebitCardsDB db, ILogger<DebitCardRepository> logger)
@@ -30,6 +31,7 @@ namespace Debit_Cards_Project.DAL.Repositories
         {
             try
             {
+                _length = _db.Cards.ToList().Count;
                 return _db.Cards.ToList();
             }
             catch (Exception e)
@@ -43,6 +45,9 @@ namespace Debit_Cards_Project.DAL.Repositories
         {
             try
             {
+                if (id < 0 || id > _length)
+                    throw new ArgumentOutOfRangeException();
+
                 return ReadAll().FirstOrDefault(card => card.Id == id)!;
             }
             catch (Exception e)
@@ -57,18 +62,23 @@ namespace Debit_Cards_Project.DAL.Repositories
             if (card is null)
                 throw new ArgumentNullException(nameof(card));
 
-            if (ReadAll().Contains(card))
-            {
-                var db_card = ReadById(id);
-                db_card.CurrencyName = card.CurrencyName;
-                db_card.Holder = card.Holder;
-                db_card.NumberCard = card.NumberCard;
-                db_card.ValidityPeriod = card.ValidityPeriod;
-            }
+            if (id < 0 || id > _length)
+                throw new ArgumentOutOfRangeException();
+
+            if (!ReadAll().Contains(card)) return;
+
+            var db_card = ReadById(id);
+            db_card.CurrencyName = card.CurrencyName;
+            db_card.Holder = card.Holder;
+            db_card.NumberCard = card.NumberCard;
+            db_card.ValidityPeriod = card.ValidityPeriod;
         }
 
         public void Delete(int id)
         {
+            if (id < 0 || id > _length)
+                throw new ArgumentOutOfRangeException();
+
             if(!ReadAll().Contains(ReadById(id))) return;
 
             _db.Cards.Remove(ReadById(id));
